Remove use of Inlining attribute
Added inlining attribute under the assumption of the byte code being so small that its always good to inline it. Reading up on how the JIT works, it probably will inline the code always anyways (or otherwise knows better what to do). So removing this as its an optimization based on assumption and not reality